# api.yml
openapi: "3.0.2"
info:
  title: Tournament manager
  version: "1.0"
servers:
  - url: http://localhost:8080/api/v1
    description: localhost
  - url: https://dev.juno.nponsard.net/api/v1
    description: dev server
  - url: https://tournament.juno.nponsard.net/api/v1
    description: prod server
paths:
...
  /users/{id}/teams:
    get:
      summary: Get teams the user participates in
      tags:
        - "users"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  admin:
                    type: boolean
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
...
  /teams:
    get:
      summary: Get all teams
      tags:
        - "teams"
      parameters:
        - in: query
          name: limit
          schema:
            type: number
          description: Number of teams to return
          required: false
        - in: query
          name: offset
          schema:
            type: number
          description: Offset of the teams to return, default 0
          required: false
        - in: query
          name: search
          schema:
            type: string
          description: Search for a team by name
          required: false
      responses:
        "200":
          description: Teams found
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: "#/components/schemas/team"
                  total:
                    type: number
                    description: Total number of teams
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    post:
      description: Create a new team, the user doing the request is automatically added to the team as captain
      tags:
        - "teams"
      security:
        - jwt: []
      requestBody:
        description: Team info
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  description:
                    type: string
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
  /teams/{id}:
    get:
      summary: Get info about a team
      tags:
        - "teams"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Team found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/team"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    patch:
      summary: Update a team info
      description: Update a team info, the user must be admin or captain of the team
      tags:
        - "teams"
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        description: Team info
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/team"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    delete:
      summary: Delete a team
      description: Delete a team, the user must be admin or captain of the team
      tags:
        - "teams"
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Team deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
  /teams/{id}/users:
    get:
      summary: Get all players of a team
      tags:
        - "teams"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Users found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
    put:
      summary: Add a player to a team
      description: Add a player to a team, the user must be admin or captain of the team
      tags:
        - "teams"
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        description: User info
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: number
                role:
                  type: number
                  description: 1 = captain, 0 = player
      responses:
        "200":
          description: User added to team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
  /teams/{id}/users/{userId}:
    delete:
      summary: Remove a player from a team
      description: Remove a player from a team, the user doing the action must be admin, captain of the team or the player to remove
      tags:
        - "teams"
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User removed from team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "500":
          $ref: "#/components/responses/500InternalServerError"
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: JWT token
  responses:
    "400BadRequest":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
    "401Unauthorized":
      description: Invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
    "403Forbidden":
      description: not enough permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
    "500InternalServerError":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/message"
  schemas:
    message:
      type: object
      description: object containing a message
      properties:
        message:
          type: string
    user:
      type: object
      description: User object
      properties:
        id:
          type: number
        username:
          type: string
    team:
      type: object
      description: Team object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        win_count:
          type: number
        match_count:
          type: number
...